bug 33483 : decimal font size
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk@6932 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ jasperreports/src/net/sf/jasperreports/components/headertoolbar/actions/EditTextElementCommand.java
@@ -23,7 +23,7 @@
  */
 package net.sf.jasperreports.components.headertoolbar.actions;
 
-import java.awt.*;
+import java.awt.Color;
 
 import net.sf.jasperreports.components.headertoolbar.HeaderToolbarElementUtils;
 import net.sf.jasperreports.components.table.util.TableUtil;
@@ -92,7 +92,7 @@ public class EditTextElementCommand implements Command
 		}
 		
 		textElement.setFontName(textElementData.getFontName());
-		textElement.setFontSize(textElementData.getFontSize() != null ? Integer.valueOf(textElementData.getFontSize()) : null);
+		textElement.setFontSize(textElementData.getFontSize() != null ? Float.valueOf(textElementData.getFontSize()) : null);
 		textElement.setBold(textElementData.getFontBold());
 		textElement.setItalic(textElementData.getFontItalic());
 		textElement.setUnderline(textElementData.getFontUnderline());
